Added NPCs shading systems and changed player's 3D model.
I got a model from the asset store and used it for the player (will probably still be a temporary one). I also made the shading system for NPCs and it's accompanying UI. Now when they walk directly under the sun the heat bar increases. It decreases if they walk in a shaded area.